Se limpia proyecto
