--- app/src/main/java/com/example/evidenciafinal/Fragment1.java
@@ -1,32 +1,0 @@
-package com.example.evidenciafinal;
-
-import android.os.Bundle;
-
-import androidx.fragment.app.Fragment;
-
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.Button;
-import android.widget.EditText;
-import android.widget.ImageButton;
-import android.widget.ImageView;
-
-
-public class Fragment1 extends Fragment {
-
-
-    public Fragment1() {
-        // Required empty public constructor
-    }
-
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_1, container, false);
-        //Para Regresar al view home
-    }
-
-}
