--- app/src/main/java/com/example/evidenciafinal/Fragment3.java
@@ -1,26 +1,0 @@
-package com.example.evidenciafinal;
-
-import android.os.Bundle;
-
-import androidx.fragment.app.Fragment;
-
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-
-public class Fragment3 extends Fragment {
-
-
-    public Fragment3() {
-        // Required empty public constructor
-    }
-
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_3, container, false);
-    }
-}
