--- app/src/main/res/layout/activity_calculos.xml
@@ -1,179 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@drawable/background3"
-    tools:context=".Calculos">
-
-    <ImageButton
-        android:id="@+id/ibBack"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="45dp"
-        android:layout_marginEnd="310dp"
-        android:layout_marginBottom="50dp"
-        android:backgroundTint="#000000"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.26"
-        app:layout_constraintStart_toStartOf="parent"
-        app:srcCompat="@android:drawable/ic_menu_revert"
-        android:contentDescription="@string/todo" />
-
-
-    <ImageButton
-        android:id="@+id/ibCalula"
-        android:layout_width="251dp"
-        android:layout_height="68dp"
-        android:layout_marginStart="80dp"
-        android:layout_marginTop="268dp"
-        android:layout_marginEnd="80dp"
-        android:background="#000000"
-        android:contentDescription="@string/todo"
-        android:scaleType="fitCenter"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:srcCompat="@drawable/boton2" />
-
-    <TextView
-        android:id="@+id/textView6"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="129dp"
-        android:layout_marginTop="140dp"
-        android:layout_marginEnd="130dp"
-        android:text="@string/indice_de_masa_corporal"
-        android:textColor="#FFFFFF"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/ibCalula" />
-
-    <TextView
-        android:id="@+id/textView7"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="170dp"
-        android:layout_marginTop="20dp"
-        android:layout_marginEnd="179dp"
-        android:text="@string/tu_imc_es"
-        android:textColor="#FFFFFF"
-        android:textSize="22sp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/textView6" />
-
-    <TextView
-        android:id="@+id/textView8"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="166dp"
-        android:layout_marginTop="8dp"
-        android:layout_marginEnd="170dp"
-        android:fontFamily="monospace"
-        android:text="@string/saludable"
-        android:textColor="#FFFFFF"
-        android:textSize="31sp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/textView7" />
-
-    <ImageButton
-        android:id="@+id/btOtrosCalculos"
-        android:layout_width="184dp"
-        android:layout_height="66dp"
-        android:layout_marginStart="21dp"
-        android:layout_marginTop="50dp"
-        android:layout_marginEnd="105dp"
-        android:layout_marginBottom="25dp"
-        android:background="#000000"
-        android:contentDescription="@string/todo"
-        android:scaleType="fitCenter"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toEndOf="@+id/ibBack"
-        app:layout_constraintTop_toBottomOf="@+id/textView8"
-        app:srcCompat="@drawable/boton4" />
-
-    <TextView
-        android:id="@+id/textView"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="93dp"
-        android:layout_marginTop="130dp"
-        android:fontFamily="monospace"
-        android:text="@string/tu_peso"
-        android:textColor="#FFFFFF"
-        android:textSize="20sp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <TextView
-        android:id="@+id/textView2"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="93dp"
-        android:layout_marginTop="7dp"
-        android:fontFamily="monospace"
-        android:text="@string/tu_altura"
-        android:textColor="#FFFFFF"
-        android:textSize="20sp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/textView" />
-
-    <TextView
-        android:id="@+id/textView3"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="93dp"
-        android:layout_marginTop="11dp"
-        android:fontFamily="monospace"
-        android:text="@string/tu_edad"
-        android:textColor="#FFFFFF"
-        android:textSize="20sp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/textView2" />
-
-    <EditText
-        android:inputType="numberDecimal"
-        android:id="@+id/etPeso"
-        android:layout_width="69dp"
-        android:layout_height="24dp"
-        android:layout_marginStart="39dp"
-        android:layout_marginTop="130dp"
-        android:background="#FFFFFF"
-        app:layout_constraintStart_toEndOf="@+id/textView"
-        app:layout_constraintTop_toTopOf="parent"
-        android:hint="@string/kg"/>
-
-    <EditText
-        android:id="@+id/etAltura"
-        android:layout_width="88dp"
-        android:layout_height="23dp"
-        android:layout_marginStart="15dp"
-        android:layout_marginTop="10dp"
-        android:background="#FFFFFF"
-        android:ems="10"
-        android:hint="m"
-        android:inputType="numberDecimal"
-        app:layout_constraintStart_toEndOf="@+id/textView2"
-        app:layout_constraintTop_toBottomOf="@+id/etPeso" />
-
-    <EditText
-        android:id="@+id/etEdad"
-        android:layout_width="88dp"
-        android:layout_height="23dp"
-        android:layout_marginStart="39dp"
-        android:layout_marginTop="15dp"
-        android:background="#FFFFFF"
-        android:ems="10"
-        android:hint="@string/a_os"
-        android:inputType="number"
-        app:layout_constraintStart_toEndOf="@+id/textView3"
-        app:layout_constraintTop_toBottomOf="@+id/etAltura"
-        android:importantForAutofill="no" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
