--- app/src/main/res/layout/activity_calorias.xml
@@ -1,151 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".Calorias"
-    android:background="@drawable/background6">
-
-
-    <TextView
-        android:id="@+id/textView2"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="77dp"
-        android:layout_marginTop="220dp"
-        android:fontFamily="monospace"
-        android:text="Kcal totales:"
-        android:textColor="#FFFFFF"
-        android:textSize="20sp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-
-    <TextView
-        android:id="@+id/textView3"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="42dp"
-        android:layout_marginTop="38dp"
-        android:fontFamily="monospace"
-        android:text="Kcal Acumuladas:"
-        android:textColor="#FFFFFF"
-        android:textSize="20sp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/btGuardarKcal" />
-
-    <EditText
-        android:id="@+id/etKcalTotal"
-        android:layout_width="88dp"
-        android:layout_height="23dp"
-        android:layout_marginStart="15dp"
-        android:layout_marginTop="224dp"
-        android:background="@color/black"
-        android:ems="10"
-        android:hint="   kcal"
-        android:inputType="number"
-        android:textColorHint="#FFFFFF"
-        android:textColor="@color/white"
-        app:layout_constraintStart_toEndOf="@+id/textView2"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <Button
-        android:id="@+id/btGuardarKcal"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="153dp"
-        android:layout_marginTop="20dp"
-        android:layout_marginEnd="153dp"
-        android:text="Guardar"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/textView2"
-        android:backgroundTint="#F6A9D1"
-        android:textColor="@color/black"/>
-
-    <EditText
-        android:id="@+id/kcalIngreso"
-        android:layout_width="88dp"
-        android:layout_height="23dp"
-        android:layout_marginStart="16dp"
-        android:layout_marginTop="40dp"
-        android:background="#FFFFFF"
-        android:ems="10"
-        android:hint="   kcal"
-        android:importantForAutofill="no"
-        android:inputType="number"
-        app:layout_constraintStart_toEndOf="@+id/textView3"
-        app:layout_constraintTop_toBottomOf="@+id/btGuardarKcal" />
-
-    <ImageButton
-        android:id="@+id/btCalcula"
-        android:layout_width="251dp"
-        android:layout_height="68dp"
-        android:layout_marginStart="80dp"
-        android:layout_marginTop="18dp"
-        android:layout_marginEnd="80dp"
-        android:background="#000000"
-        android:contentDescription="@string/todo"
-        android:scaleType="fitCenter"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/tvRestantes"
-        app:srcCompat="@drawable/boton2" />
-
-    <ImageButton
-        android:id="@+id/ibBack"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="351dp"
-        android:layout_marginEnd="4dp"
-        android:layout_marginBottom="16dp"
-        android:backgroundTint="#000000"
-        android:contentDescription="@string/todo"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:srcCompat="@android:drawable/ic_menu_revert" />
-
-    <TextView
-        android:id="@+id/tvRestantes"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="66dp"
-        android:layout_marginTop="28dp"
-        android:fontFamily="monospace"
-        android:text="Kcal a restar:"
-        android:textColor="#FFFFFF"
-        android:textSize="20sp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/textView3" />
-
-    <EditText
-        android:id="@+id/etKcalRestantes"
-        android:layout_width="88dp"
-        android:layout_height="23dp"
-        android:layout_marginStart="12dp"
-        android:layout_marginTop="32dp"
-        android:ems="10"
-        android:hint="   - kcal"
-        android:inputType="number"
-        android:background="@color/white"
-        app:layout_constraintStart_toEndOf="@+id/tvRestantes"
-        app:layout_constraintTop_toBottomOf="@+id/kcalIngreso" />
-
-    <Button
-        android:id="@+id/btGuardarDB"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="198dp"
-        android:layout_marginTop="81dp"
-        android:layout_marginEnd="92dp"
-        android:backgroundTint="#3E1E77"
-        android:text="Fin del día"
-        android:textColor="@color/white"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/btCalcula" />
-
-
-</androidx.constraintlayout.widget.ConstraintLayout>
